--- a/aws-lc-rs/Cargo.toml
+++ b/aws-lc-rs/Cargo.toml
@@ -31,7 +31,7 @@ fips = ["dep:aws-lc-fips-sys"]
 [dependencies]
 untrusted = { version = "0.7.1", optional = true }
 aws-lc-sys = { version = "0.6.1", path = "../aws-lc-sys", optional = true }
-aws-lc-fips-sys = { version = "0.7.1", path = "../aws-lc-fips-sys", optional = true }
+aws-lc-fips-sys = { version = "0.8.0", path = "../aws-lc-fips-sys", optional = true }
 zeroize = "1"
 mirai-annotations = "1.12.0"