### Issues:
N/A

### Description of changes: 
* Update aws-lc-fips-sys to version 0.8.0
* aws-lc-fips-sys no longer requires a C++ compiler to build.
* Improved build times:

Before:
```
❯ time cargo build 
...
   Compiling aws-lc-fips-sys v0.7.1 (/home/justsmth/repos/aws-lc-rs/aws-lc-fips-sys)
    Finished dev [unoptimized + debuginfo] target(s) in 16.61s
cargo build  61.70s user 8.98s system 424% cpu 16.645 total
```

After:
```
❯ time cargo build
...
   Compiling aws-lc-fips-sys v0.8.0 (/home/justsmth/repos/aws-lc-rs/aws-lc-fips-sys)
    Finished dev [unoptimized + debuginfo] target(s) in 11.77s
cargo build  32.63s user 5.87s system 326% cpu 11.803 total

```

### Call-outs:
N/A

### Testing:
Crate generation succeeded.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license and 
the ISC license.